requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
constantly==15.1.0
cryptography==37.0.4
cssselect2==0.6.0
daphne==3.0.2
decorator==4.4.2
Deprecated==1.2.13
Django==4.0.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-eventstream==4.4.1
django-filer==2.2.3
django-filter==22.1
django-grip==3.2.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-mptt==0.13.4
django-notifications==0.1.dev0
django-notifications-hq==1.7.0
django-notifications-rest==0.0.1
django-polymorphic==3.1.0
django-redis==5.2.0
django-redis-sessions==0.6.2
django-timezone-field==5.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
easy-thumbnails==2.8.3
flake8==5.0.4
gripcontrol==4.1.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
incremental==21.3.0
iniconfig==1.1.1
jsonfield==3.1.0
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
moviepy==1.0.3
msgpack==1.0.4
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
proglog==0.1.10
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pubcontrol==3.3.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-sugar==0.9.5
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.2.1
redgrease==0.3.30
redis==4.3.4
reportlab==3.6.11
requests==2.28.1
rq==1.11.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
swapper==1.3.0
termcolor==1.1.0
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
Unidecode==1.1.2
urllib3==1.26.12
uvicorn==0.18.3
vine==5.0.0
walrus==0.9.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.4.0
